Bug - Fix Pyright's `reportUnitializedClassVariable` error & switch to strict (#69)
This PR introduces strict type checking (except a couple rules) for all
the repo. I fixed type issues, removed `mypy` because it conflicts on
it's type checking for async functions from Pyright which is impossible
to handle, and added a `@t.dataclass_transform()` to `Command` so that
it's interpreted as a dataclass which lets users have non-default args.