Bug - Fix field validation (#79)
In https://github.com/danimelchor/clypi/pull/76 I moved part of the
parsing logic to `__init__` since it's called to parse commands.
However, this breaks forwarded arguments since we're not failing as
early as we were before.

For example, if a forwarded argument is missing, we're now doing it
during class instantiation. This means that we're doing it **after** we
parse the subcommands, so the error displayed actually comes from the
forwarded field missing and not from the parent command which required
the field in the first place.