Feat - Display forwarded arguments (#81)
Allow forwarding options with all of their configs to the child commands
and adds a new option to show them in the help formatter. If a forwarded
option is now passed to a child subcommand, the parent will grab the
option from it. Therefore, the configs are forwarded to child
subcommands, and the values are fetched upwards from the childrent to
the parent commands